LayoutTests/accessibility/gtk/html-slider-indicator-expected.txt


This tests that a basic range returns all the correct information for the gtk platform.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS range.minValue is 0
PASS range.maxValue is 100
PASS range.intValue is 50
PASS range.intValue is 49
PASS range.intValue is 50
PASS range.intValue is 55
PASS range.intValue is 50
PASS range.isEqual(hitTestRange) is true
PASS successfullyParsed is true

TEST COMPLETE
Successfully received AXValueChanged
Successfully received AXValueChanged
Successfully received AXValueChanged
Successfully received AXValueChanged
